pyproject.toml
[project]
name = "litellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.26.0",
    "ipykernel>=6.29.5",
    "litellm>=1.71.3",
    "ollama>=0.5.1",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
]
